backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt==3.2.2",
    "beautifulsoup4>=4.14.3",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.127.1",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "mem0ai>=1.0.1",
    "oci>=2.164.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.5.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "rq>=2.6.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
    "sse-starlette>=3.1.1",
    "tavily-python>=0.7.17",
]
